LayoutTests/platform/ios-simulator/editing/pasteboard/data-transfer-items-drag-drop-entry-expected.txt

CONSOLE MESSAGE: line 56: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(files)')
FAIL: Timed out waiting for notifyDone to be called
This tests the basic functionality and properties of DataTransferItems for files with drag and drop. This test requires DRT.
Drop files here if you test this manually
Dragging file: resources/mozilla.gif
Dragging file: resources/drop-file-svg.svg
Dragging file: resources/copy-backslash-euc.html
Dragging file: resources/test_directory
